backend/requirements.txt
fastapi==0.100.0
uvicorn==0.15.0
langchain==0.0.267
openai==0.27.0
websockets==10.4
python-dotenv==0.21.0
httpx>=0.24.0
pydantic-settings==2.0.2
pandas==1.3.3
pydantic==2.0.2
SQLAlchemy==2.0.34
pgvector==0.3.2
psycopg2-binary==2.9.9
Pillow==10.0.0
rembg==2.0.50
python-multipart
